+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Changed
+* `parsing.ErrorResponse` fields `message` and `detail` are now private,
+  encapsulated with getters.
+
+
 ## [0.2.1] - 2022-10-19
 ### Fixed
 * Handle case where HTTP response is not valid JSON.
@@ -45,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-java/compare/v0.2.1...HEAD
 [0.2.1]: https://github.com/DeepLcom/deepl-java/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/DeepLcom/deepl-java/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/DeepLcom/deepl-java/compare/v0.1.2...v0.1.3